consolidate post & delete epics
