+++ src/middleware/epic.js
@@ -3,8 +3,10 @@ import { combineEpics, createEpicMiddleware } from 'redux-observable';
 
 import getSyncEpic from '../epics/sync';
 import getCacheEpic from '../epics/cache';
-import getPostEpic from '../epics/post';
-import getDeleteEpic from '../epics/delete';
+import {
+	getPostEpic,
+	getDeleteEpic
+} from '../epics/mutate';
 
 /**
  * The middleware is exported as a getter because it needs the application's
